% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkers.R
\name{check_adae}
\alias{check_adae}
\title{ADAE dataset checker}
\usage{
check_adae(adam)
}
\arguments{
\item{adam}{A dataframe that follows the CDISC Analysis Data Model (ADaM)
standards.}
}
\value{
A single \code{logical}, \code{TRUE} if the check passes and \code{FALSE} if not, with attributes:
\itemize{
\item \code{msg}: \code{character} vector of reasons for check failure
\item \code{data}: \code{adam}
}
}
\description{
ADAE dataset checker
}
\examples{
dplyr::tibble(
  STUDYID = character("1"),
  USUBJID = character("1"),
  AEDECOD = character("1"),
  AEBODSYS = character("1"),
  AEREFID = character("1"),
  AECYCVID = character("1"),
  AESER = character("1"),
  ASTDT = integer(1L)
) |>
  check_adae()

}
