% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkers.R
\name{check_has_any_cols}
\alias{check_has_any_cols}
\title{Check if a group of columns exist in a dataset}
\usage{
check_has_any_cols(column_names)
}
\arguments{
\item{column_names}{A character vector of column names expected in the
dataset.}
}
\value{
Checker function that takes an ADaM dataset as argument and returns a \code{logical} check result.
See \code{?check_adam} for example.
}
\description{
Check if a group of columns exist in a dataset
}
\examples{
param_col_names <- c("PARAMCD", "PARAM", "PARAMN")
dplyr::tibble(
  PARAMCD = "A", PARAM = "B") |>
  check_has_any_cols(column_names = param_col_names)()

param_col_names <- c("PARAMCD", "PARAM", "PARAMN")
dplyr::tibble(
  A = "A") |>
  check_has_any_cols(column_names = param_col_names)()

}
