% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkers.R
\name{check_has_cols}
\alias{check_has_cols}
\title{Check if columns exist in a dataset}
\usage{
check_has_cols(column_names)
}
\arguments{
\item{column_names}{A character vector of column names expected in the
dataset.}
}
\value{
Checker function that takes an ADaM dataset as argument and returns a \code{logical} check result.
See \code{?check_adam} for example.
}
\description{
Check if columns exist in a dataset
}
\examples{
dplyr::tibble(col_a = 1, col_b = 2) |>
  check_has_cols(column_names = c("col_a", "col_b"))()

}
