% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkers.R
\name{check_has_one_record_per_subject_per_param}
\alias{check_has_one_record_per_subject_per_param}
\title{Check if dataset has one record per subject per parameter}
\usage{
check_has_one_record_per_subject_per_param(adam)
}
\arguments{
\item{adam}{A dataframe that follows the CDISC Analysis Data Model (ADaM)
standards.}
}
\value{
A single \code{logical}, \code{TRUE} if the check passes and \code{FALSE} if not, with attributes:
\itemize{
\item \code{msg}: \code{character} vector of reasons for check failure
\item \code{data}: \code{adam}
}
}
\description{
Check if dataset has one record per subject per parameter
}
\examples{
dplyr::tibble(STUDYID = 1, USUBJID = 1, PARAMCD = 1) |>
  check_has_one_record_per_subject_per_param()

}
