% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkers.R
\name{check_is_categorical}
\alias{check_is_categorical}
\title{Check if a column is a factor category}
\usage{
check_is_categorical(column_name)
}
\arguments{
\item{column_name}{The column name to check datatype.}
}
\value{
Checker function that takes an ADaM dataset as argument and returns a \code{logical} check result.
See \code{?check_adam} for example.
}
\description{
Check if a column is a factor category
}
\examples{
sex <- factor(c("male, female"))

dplyr::tibble(
  sex
) |> check_is_categorical("sex")()

sex <- c("male, female")

dplyr::tibble(
  sex
) |> check_is_categorical("sex")()

}
