% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkers.R
\name{check_is_character}
\alias{check_is_character}
\title{Check if a column is character}
\usage{
check_is_character(column_name)
}
\arguments{
\item{column_name}{The column name to check datatype.}
}
\value{
Checker function that takes an ADaM dataset as argument and returns a \code{logical} check result.
See \code{?check_adam} for example.
}
\description{
Check if a column is character
}
\examples{
comments <- "text"

dplyr::tibble(
  comments
) |> check_is_character("comments")()

comments <- factor(c("male, female"))

dplyr::tibble(
  comments
) |> check_is_character("comments")()

}
