% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkers.R
\name{check_is_date}
\alias{check_is_date}
\title{Check if a column is date}
\usage{
check_is_date(column_name)
}
\arguments{
\item{column_name}{The column name to check datatype.}
}
\value{
Checker function that takes an ADaM dataset as argument and returns a \code{logical} check result.
See \code{?check_adam} for example.
}
\description{
Check if a column is date
}
\examples{
library(lubridate)
test_date <- as.Date("2003-02-27")
dplyr::tibble(
  test_date
) |> check_is_date("test_date")()

test_date <- as.POSIXct("2003-02-27 22:22:22")

dplyr::tibble(
  test_date
) |> check_is_date("test_date")()

}
