% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkers.R
\name{check_is_datetime}
\alias{check_is_datetime}
\title{Check if a column is datetime}
\usage{
check_is_datetime(column_name)
}
\arguments{
\item{column_name}{The column name to check datatype.}
}
\value{
Checker function that takes an ADaM dataset as argument and returns a \code{logical} check result.
See \code{?check_adam} for example.
}
\description{
Check if a column is datetime
}
\examples{
library(lubridate)
test_datetime <- as.POSIXct("2020/01/01 22:22:22")
dplyr::tibble(
  test_datetime
) |> check_is_datetime("test_datetime")()

test_datetime <- as.Date("2020/01/01")

dplyr::tibble(
  test_datetime
) |> check_is_datetime("test_datetime")()

}
