% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkers.R
\name{check_is_double}
\alias{check_is_double}
\title{Check if a column is double}
\usage{
check_is_double(column_name)
}
\arguments{
\item{column_name}{The column name to check datatype.}
}
\value{
Checker function that takes an ADaM dataset as argument and returns a \code{logical} check result.
See \code{?check_adam} for example.
}
\description{
Check if a column is double
}
\examples{
USUBJID <- 1.1
dplyr::tibble(
  USUBJID
) |> check_is_double("USUBJID")()

test_datetime <- as.POSIXct("2020/01/01 22:22:22")

dplyr::tibble(
  test_datetime
) |> check_is_double("test_datetime")()

}
