% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkers.R
\name{check_is_integerish}
\alias{check_is_integerish}
\title{Check if a column is integerish}
\usage{
check_is_integerish(column_name)
}
\arguments{
\item{column_name}{The column name to check datatype.}
}
\value{
Checker function that takes an ADaM dataset as argument and returns a \code{logical} check result.
See \code{?check_adam} for example.
}
\description{
Check if a column is integerish
}
\examples{
USUBJID <- as.integer(1.1)
dplyr::tibble(
  USUBJID
) |> check_is_integerish("USUBJID")()

USUBJID <- as.double(1)

dplyr::tibble(
  USUBJID
) |> check_is_integerish("USUBJID")()

USUBJID <- as.double(1.1)

dplyr::tibble(
  USUBJID
) |> check_is_integerish("USUBJID")()

}
