% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkers.R
\name{check_is_logical}
\alias{check_is_logical}
\title{Check if a column is logical}
\usage{
check_is_logical(column_name)
}
\arguments{
\item{column_name}{The column name to check datatype.}
}
\value{
Checker function that takes an ADaM dataset as argument and returns a \code{logical} check result.
See \code{?check_adam} for example.
}
\description{
Check if a column is logical
}
\examples{
status <- TRUE

dplyr::tibble(
  status
) |> check_is_logical("status")()

x <- c("male, female")
status <- factor(x)

dplyr::tibble(
  status
) |> check_is_logical("status")()

}
