% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkers.R
\name{check_lacks_values}
\alias{check_lacks_values}
\title{Check if a column lacks specific values}
\usage{
check_lacks_values(column_name, values)
}
\arguments{
\item{column_name}{The column name to check values of.}

\item{values}{A vector of values expected to not be present in the specified
column.}
}
\value{
Checker function that takes an ADaM dataset as argument and returns a \code{logical} check result.
See \code{?check_adam} for example.
}
\description{
Check if a column lacks specific values
}
\examples{
dplyr::tibble(col_a = 1) |>
check_lacks_values(column_name = "col_a", values = 2)()

}
