% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_checkers.R
\name{make_checker}
\alias{make_checker}
\title{Create a checker function}
\usage{
make_checker(fn)
}
\arguments{
\item{fn}{A function that accepts an ADaM dataset and returns a named list
with two elements: 'check' - a logical value that indicates if the ADaM
dataset passed the check, and 'msg' - a character string that provides a
description of the failure if the check did not pass.}
}
\value{
Checker function that takes an ADaM dataset as argument and returns a \code{logical} check result.
See \code{?check_adam} for example.
}
\description{
Construct a function that performs a specific check on an ADaM dataset.
}
\seealso{
\code{make_multi_checker} for combining checker functions into one.
}
