% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_checkers.R
\name{make_multi_checker}
\alias{make_multi_checker}
\title{Create a function for multiple checks}
\usage{
make_multi_checker(...)
}
\arguments{
\item{...}{One or more checker functions created by \code{make_checker}.}
}
\value{
Checker function that takes an ADaM dataset as argument and returns a \code{logical} check result.
See \code{?check_adam} for example.
}
\description{
Combine one or more checker functions into one.
}
\details{
Checks ran by functions passed to \code{...} are performed sequentially and check results are aggregated.
}
\seealso{
\code{make_checker} for creating individual checker functions.
}
