% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teal.builder.checks-package.R
\docType{package}
\name{teal.builder.checks-package}
\alias{teal.builder.checks}
\alias{teal.builder.checks-package}
\title{teal.builder.checks: Data Compatibility Checks for Teal Builder Modules}
\description{
These functions assess the compatibility of ADaM datasets with particular Teal Builder modules. If a dataset is incompatible, they generate descriptive messages explaining the reasons for the incompatibility.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Sanofi-Public/teal.builder.checks}
  \item Report bugs at \url{https://github.com/Sanofi-Public/teal.builder.checks/issues}
}

}
\author{
\strong{Maintainer}: Edward Gillian \email{edward.gillian-ext@sanofi.com}

Other contributors:
\itemize{
  \item Sanofi [copyright holder, funder]
}

}
\keyword{internal}
